requirements.txt
aiogram==3.14.0
aiohttp==3.11.11
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
httpx==0.27.2
apscheduler==3.10.4
pydantic==2.10.3
pydantic-settings==2.7.0
certifi==2024.8.30